Improve project filtering on Upcoming tab & add delete project confirmation
- Fixed the project filter on the "Upcoming" tab so it now correctly displays tasks only for the selected project.

- Added a confirmation warning before deleting a project to prevent accidental data loss.

- Refactored the \`renderUpcomingTasks\` function to accept a task list. This change was necessary to let the project filtering on the Upcoming tab work.

+++ src/handlers/projects/addProjectMenu.js
@@ -74,14 +74,21 @@ export default function addProjectMenu() {
   }
   if (deleteProjectBtn) {
     deleteProjectBtn.addEventListener("click", () => {
-      const deleteProjectID = projectModal.dataset.id;
-      if (!deleteProjectID) {
-        console.error("No project ID set in modal. Cannot delete.");
+      const userInput = prompt(
+        "Deleting a project removes it from all attached tasks: Are you sure ? y/N"
+      );
+      if (userInput === "y") {
+        const deleteProjectID = projectModal.dataset.id;
+        if (!deleteProjectID) {
+          console.error("No project ID set in modal. Cannot delete.");
+          return;
+        }
+        deleteProject(deleteProjectID);
+        renderAllProjects();
+        projectModal.close();
+      } else {
         return;
       }
-      deleteProject(deleteProjectID);
-      renderAllProjects();
-      projectModal.close();
     });
   }
   if (deleteAllProjectsBtn) {
